Create README.md
This file is just a variation on the Ship Station Plotter in my repository, but works with Amazon sales data.  

A limitation of this file is that it only can take ytd information.  If you need prior years' info, you'll need to run it twice, three times, depending on however many years you are going back.

Basically it takes sales information per product, and charts the performance of the top n products by volume.  The "0" column is the Coefficient, or the slope of the sales performance line, which you can sort, and thereby rank, products by the acceleration of their sales velocity.  

The chart works in the Ship Station Plotter.  I think I messed that up while adapting this file to suit Amazon data.